Create base numerical column class for numerical and decimal.
